requirements.txt
Flask==2.3.3
Flask-PyMongo==2.3.0
pymongo==4.5.0
Flask-JWT-Extended==4.5.2
Flask-Bcrypt==1.0.1
Flask-Cors==4.0.0
stripe==6.7.0
python-dotenv==1.0.0
email-validator==2.0.0
gunicorn==21.2.0
google-api-python-client==2.99.0
google-auth==2.23.0
google-auth-httplib2==0.1.0
requests==2.31.0